修复部分小数格式问题

+++ watermeter/src/main/java/com/android/blm/watermeter/bean/Bean_RechargeRecord.java
@@ -16,7 +16,7 @@ public class Bean_RechargeRecord {
         public String OperID;
         public String OperDate;
         public String LastBalance;
-        public float PayMoney;
+        public double PayMoney;
         public String ThisBalance;
     }
 }
